Add files via upload
The hyperplane (which in two dimensions is simply a line) that best separates the tags is produced by a support vector machine using these data points. This line serves as a decision boundary: anything on one side will be classified as cyan, and anything on the other will be classified as red.